Added line tracking
Code now includes functions to write the current ramp line number to
the SD card each time the ramp progresses (function updateRampPos) and
retrieve this number when the Arduino starts up (function getRampLine).
These allow the Arduino to continue running the same ramp, instead of
restarting it, in the event of a reset during power outages.